Continue working on bbox
